Two useful naming idioms are found in this commi:
1. As in Prolog, use the identifier '_' for anything which is unworthy
of a name. That is, for anything which the bloody language rules require
you to identify by name, even though you are never going to use this
name. Unforunately, unlike Prolog, Java does not allow you to use the
'_' identifier, nor any other identifier, for different purposes in the
same scope.If you encounter two or more unworthy of a name thingies,
cosnider using identifiers such as '__', '____' , etc. 

2. As in Regular expressions, use the identifier '$' (surprise,
surprise, '$' is a perfectly valid identifier in Java) for denoting the
end, the goal, the result of product of the current computation. It is
typical that a function defines a variable, and then carries out its
a bit of computation to produce the result in that variable. the
convention proposed is to name this variable '$'. 

Note that in our example, we extend this conventio to include parametes.
If a function returns that its parameter, perhaps after conversion into
a differentt format or type, then the idiom suggests: name this
parameter '$'