Check if 0 is returned when no methods are present #679
A new test that makes sure 0 is returned when a compilation unit with no methods has been added. This test didn't require the change of the 'methods' function since it already returned 0, and all the tests passed.
In addition, a helper function has been added to the test class to easily create compilation units from strings.