Marked code#essence as deprecated #833
Due to conceptual duplication between Essence#of and code#essence, code#essence was marked as deprecated.
All calls to code#essence were replaced with calls to Essence#of.
Suppressed warnings for "deprecation" in our test class.
Added relevant documentation to code#essence.
Since the function is now deprecated, the TODO inside it was removed.
All tests are passing.